+++ Tests/AppTests/PackageController+ShowRouteTests.swift
@@ -47,7 +47,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         try await Version(package: pkg, reference: .tag(.init(2, 0, 0, "beta2"), "2.0.0beta2")).create(on: app.db)
 
         // MUT
-        let record = try await API.PackageController.GetRoute.History.query(on: app.db, owner: "foo", repository: "bar").unwrap()
+        let record = try await API.PackageController.History.query(on: app.db, owner: "foo", repository: "bar").unwrap()
 
         // validate
         XCTAssertEqual(
@@ -74,7 +74,7 @@ class PackageController_ShowRouteTests: AppTestCase {
                              owner: "foo").create(on: app.db)
 
         // MUT
-        let record = try await API.PackageController.GetRoute.History.query(on: app.db, owner: "foo", repository: "bar").unwrap()
+        let record = try await API.PackageController.History.query(on: app.db, owner: "foo", repository: "bar").unwrap()
 
         // validate
         XCTAssertEqual(
@@ -91,7 +91,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         Current.date = { .spiBirthday }
         do {  // all inputs set to non-nil values
             // setup
-            let record = API.PackageController.GetRoute.History.Record(
+            let record = API.PackageController.History.Record(
                 url: "url",
                 defaultBranch: "main",
                 firstCommitDate: .t0,
@@ -114,7 +114,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         }
         do {  // test nil inputs
             XCTAssertNil(
-                API.PackageController.GetRoute.History.Record(
+                API.PackageController.History.Record(
                     url: "url",
                     defaultBranch: nil,
                     firstCommitDate: .t0,
@@ -123,7 +123,7 @@ class PackageController_ShowRouteTests: AppTestCase {
                 ).historyModel()
             )
             XCTAssertNil(
-                API.PackageController.GetRoute.History.Record(
+                API.PackageController.History.Record(
                     url: "url",
                     defaultBranch: "main",
                     firstCommitDate: nil,
@@ -163,7 +163,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         }
 
         // MUT
-        let res = try await API.PackageController.GetRoute.ProductCount.query(on: app.db, owner: "foo", repository: "bar")
+        let res = try await API.PackageController.ProductCount.query(on: app.db, owner: "foo", repository: "bar")
 
         // validate
         XCTAssertEqual(res.filter(\.isExecutable).count, 1)
@@ -241,7 +241,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         ]
 
         // MUT
-        let res = API.PackageController.GetRoute.BuildInfo
+        let res = API.PackageController.BuildInfo
             .platformBuildResults(builds: builds, kind: .defaultBranch)
 
         // validate
@@ -276,7 +276,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         ]
 
         // MUT
-        let res = API.PackageController.GetRoute.BuildInfo
+        let res = API.PackageController.BuildInfo
             .swiftVersionBuildResults(builds: builds, kind: .defaultBranch)
 
         // validate
@@ -295,7 +295,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         ]
 
         // MUT
-        let res = API.PackageController.GetRoute.BuildInfo.platformBuildInfo(builds: builds)
+        let res = API.PackageController.BuildInfo.platformBuildInfo(builds: builds)
 
         // validate
         XCTAssertEqual(res?.stable?.referenceName, "1.2.3")
@@ -319,7 +319,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         ]
 
         // MUT
-        let res = API.PackageController.GetRoute.BuildInfo.swiftVersionBuildInfo(builds: builds)
+        let res = API.PackageController.BuildInfo.swiftVersionBuildInfo(builds: builds)
 
         // validate
         XCTAssertEqual(res?.stable?.referenceName, "1.2.3")
@@ -380,7 +380,7 @@ class PackageController_ShowRouteTests: AppTestCase {
         }
 
         // MUT
-        let res = try await API.PackageController.GetRoute.BuildInfo.query(on: app.db, owner: "foo", repository: "bar")
+        let res = try await API.PackageController.BuildInfo.query(on: app.db, owner: "foo", repository: "bar")
 
         // validate
         // just test reference names and some details for `latest`
